requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
av==16.0.1
black==25.9.0
blessings==1.7
build==1.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
decord==0.6.0
# diffusers==0.35.2  # Installed separately in Dockerfile
dm-tree==0.1.9
docstring-parser==0.17.0
draccus==0.11.5
einops==0.8.1
etils==1.13.0
farama-notifications==0.0.4
fastparquet==2024.11.0
filelock==3.20.0
# flash-attn @ file:///media/deepansh/44781eab-8a50-4dc1-942e-29a9a59d4de6/humanoid_ws/Isaac-GR00T/flash-attention
fonttools==4.60.1
fsspec==2025.9.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
gitpython==3.1.45
glfw==2.10.0
# -e file:///media/deepansh/44781eab-8a50-4dc1-942e-29a9a59d4de6/humanoid_ws/src/Isaac-GR00T
grpcio==1.76.0
gymnasium==1.2.1
h5py==3.15.1
hf-xet==1.1.10
# huggingface-hub==0.35.3  # Installed separately in Dockerfile
hydra-core==1.3.2
idna==3.11
imageio==2.37.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.3.0
iopath==0.1.10
isort==7.0.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
# kornia==0.8.1  # Installed separately in Dockerfile
# kornia-rs==0.1.9  # Installed separately in Dockerfile
llvmlite==0.45.1
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.2
# matplotlib==3.10.7  # Installed separately in Dockerfile
mdurl==0.1.2
mergedeep==1.3.4
ml-dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.3.7
mypy==1.18.2
mypy-extensions==1.1.0
networkx==3.5
numba==0.62.1
# numpy==2.3.4  # Installed separately in Dockerfile
numpydantic==1.7.0
# nvidia-cublas-cu12==12.8.4.1  # Comes with torch
# nvidia-cuda-cupti-cu12==12.8.90  # Comes with torch
# nvidia-cuda-nvrtc-cu12==12.8.93  # Comes with torch
# nvidia-cuda-runtime-cu12==12.8.90  # Comes with torch
# nvidia-cudnn-cu12==9.10.2.21  # Comes with torch
# nvidia-cufft-cu12==11.3.3.83  # Comes with torch
# nvidia-cufile-cu12==1.13.1.3  # Comes with torch
# nvidia-curand-cu12==10.3.9.90  # Comes with torch
# nvidia-cusolver-cu12==11.7.3.90  # Comes with torch
# nvidia-cusparse-cu12==12.5.8.93  # Comes with torch
# nvidia-cusparselt-cu12==0.7.1  # Comes with torch
# nvidia-nccl-cu12==2.27.5  # Comes with torch
# nvidia-nvjitlink-cu12==12.8.93  # Comes with torch
# nvidia-nvshmem-cu12==3.3.24  # Comes with torch
# nvidia-nvtx-cu12==12.8.90  # Comes with torch
omegaconf==2.3.0
onnx==1.19.1
# opencv-python==4.11.0.86  # Installed separately in Dockerfile
# opencv-python-headless==4.11.0.86  # Installed separately in Dockerfile
packaging==25.0
# pandas==2.3.3  # Installed separately in Dockerfile
pathspec==0.12.1
# peft==0.17.1  # Installed separately in Dockerfile
pettingzoo==1.25.0
# pillow==12.0.0  # Installed separately in Dockerfile
pip==25.3
pipablepytorch3d==0.7.6
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
protobuf==6.33.0
psutil==7.1.1
pyarrow==21.0.0
pydantic==2.12.3
pydantic-core==2.41.4
pygame==2.6.1
pygments==2.19.2
pyopengl==3.1.10
pyparsing==3.2.5
pyproject-hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytz==2025.2
pyyaml==6.0.3
pyyaml-include==1.4.1
pyzmq==27.1.0
ray==2.50.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-click==1.9.4
rpds-py==0.27.1
ruff==0.14.2
# safetensors==0.6.2  # Installed separately in Dockerfile
# scipy==1.16.2  # Installed separately in Dockerfile
sentry-sdk==2.42.1
setuptools==80.9.0
shtab==1.7.2
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
stringzilla==4.2.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tianshou==0.5.1
# timm==1.0.20  # Installed separately in Dockerfile
# tokenizers==0.21.4  # Installed separately in Dockerfile
toml==0.10.2
tqdm==4.67.1
# transformers==4.51.3  # Installed separately in Dockerfile
triton==3.5.0
typeguard==4.4.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tyro==0.9.35
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.2
werkzeug==3.1.3
wrapt==2.0.0
yacs==0.1.8
zipp==3.23.0